read multiple playlists

+++ playlist_html_finder.py
@@ -22,13 +22,16 @@ def parse_playlists():
     lines: list[tuple[str, str]] = []
     existing_json = os.listdir(output_folder)
     with open(playlists_filename) as f:
-        line = f.readline().strip()
-        line = line.split(' ')
-        if len(line) == 1:
-            line.append(line[0]) #let playlist id be name of output filename
-        already_exists = f'{line[1]}.json' in existing_json
-        if not already_exists:
-            lines.append(line)
+        lines_read = f.readlines()
+        for line in lines_read:
+            line = line.strip()
+            if not line: continue #line empty
+            line = line.split(' ')
+            if len(line) == 1:
+                line.append(line[0]) #let playlist id be name of output filename
+            already_exists = f'{line[1]}.json' in existing_json
+            if not already_exists:
+                lines.append(line)
 
     if not lines:
         print('No new playlists given. Delete existing in output folder if you want to redo')
@@ -38,26 +41,31 @@ def parse_playlists():
     driver.implicitly_wait(10)
 
     for playlist_id, playlist_name in lines:
-        url = f"https://www.youtube.com/playlist?list={playlist_id}"
-        print("Getting page")
-        driver.get(url)
-        print("Finding elements")
-        elem = driver.find_element_by_tag_name("html")
-        print("Scrolling")
-        elem.send_keys(Keys.END)
-        time.sleep(3)
-        elem.send_keys(Keys.END)
-        print("Getting inner html")
-        innerHTML = driver.execute_script("return document.body.innerHTML")
 
-        print("Writing html to file")
         html_folder = os.path.join(output_folder, 'html')
         if not os.path.exists(html_folder):
             os.mkdir(html_folder)
 
         html_filename = os.path.join(output_folder, 'html', f'{playlist_name}.html')
-        with open(html_filename, 'w') as f:
-            f.write(innerHTML)
+        #if html exists then dont get again
+        if not os.path.exists(html_filename):
+            url = f"https://www.youtube.com/playlist?list={playlist_id}"
+            print(f"Getting page for {playlist_name}")
+            driver.get(url)
+            print("Finding elements")
+            elem = driver.find_element_by_tag_name("html")
+            print("Scrolling")
+            elem.send_keys(Keys.END)
+            time.sleep(3)
+            elem.send_keys(Keys.END)
+            print("Getting inner html")
+            innerHTML = driver.execute_script("return document.body.innerHTML")
+
+            print("Writing html to file")
+            with open(html_filename, 'w') as f:
+                f.write(innerHTML)
+        else:
+            print(f"Using existing page for {playlist_name}")
 
         output_filename = os.path.join(output_folder, f'{playlist_name}.json')
         create_songs_json(html_filename, output_filename)
